### 2025-05-17 #8215 @finsberg
> **Submitting author:** [@finsberg](https://github.com/finsberg) ([Henrik Nicolay Topnes Finsberg](http://orcid.org/0000-0003-3766-2393)) **Repository:** https://github.com/finsberg/fenicsx-beat **Branch with paper.md** (empty if default branch): **Version:** v0.1.1 **Editor:** Pending **Reviewers:** Pending **Managing EiC:** Kevin M. Moerman
>
> ## Status
> [![status](https://camo.githubusercontent.com/eb77c498f9fcacae3c67319fd58c9234f853a2f5cf0137f17beb55d3de9c5e39/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f65633563383530663164333334646633306534633065633231326433323835302f7374617475732e737667)](https://joss.theoj.org/papers/ec5c850f1d334df30e4c0ec212d32850)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS [@finsberg](https://github.com/finsberg). **Currently, there isn't a JOSS editor assigned** to your paper.
>
> [@finsberg](https://github.com/finsberg) if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot [@editorialbot](https://github.com/editorialbot) is here to help you find and assign reviewers and start the main review. To find out what [@editorialbot](https://github.com/editorialbot) can do for you type:
>
> ```
> @editorialbot commands
> ```

I think `ashahide`, `sitic`, `cudmore`, `oalii` and `SunnyXu` could be potential reviewers.
