### 2025-05-25 1033e22
Add calculation of qt-interval
