+++ .cspell_dict.txt
@@ -4,6 +4,7 @@ allclose
 allreduce
 arange
 Arens
+argmin
 argsort
 Aslak
 astype
@@ -11,6 +12,7 @@ atol
 Augustin
 Bartolucci
 basix
+bazett
 bcast
 bestel
 Bestel
@@ -54,6 +56,7 @@ Fitz
 fitzhughnagumo
 Francesco
 Frederik
+fridericia
 functionspace
 genexpr
 geodir
@@ -130,6 +133,7 @@ preonly
 pyplot
 PYVISTA
 Quarteroni
+randn
 regazzoni
 Regazzoni
 Remedios
