### 2025-05-26 076f463
Do not transfer values between ODE and PDE, but rather use ODE directly in PDE
