### 2025-05-26 8ae1b5e
Revert "Do not transfer values between ODE and PDE, but rather use ODE directly"
