+++ tests/test_monodomain_solver.py
@@ -51,10 +51,10 @@ def test_monodomain_splitting_analytic(odespace):
 
     I_s = ac_func(x, time)
 
+    pde = beat.MonodomainModel(time=time, mesh=mesh, M=M, I_s=I_s)
+
     V_ode = beat.utils.space_from_string(odespace, mesh, dim=1)
     v_ode = dolfinx.fem.Function(V_ode)
-
-    pde = beat.MonodomainModel(time=time, mesh=mesh, M=M, I_s=I_s, v_ode=v_ode)
 
     s = dolfinx.fem.Function(V_ode)
     s.interpolate(
@@ -67,6 +67,7 @@ def test_monodomain_splitting_analytic(odespace):
 
     ode = beat.odesolver.DolfinODESolver(
         v_ode=v_ode,
+        v_pde=pde.state,
         fun=simple_ode_forward_euler,
         init_states=init_states,
         parameters=None,
@@ -108,10 +109,11 @@ def test_monodomain_splitting_spatial_convergence(odespace):
 
         I_s = ac_func(x, time)
 
+        pde = beat.MonodomainModel(time=time, mesh=mesh, M=M, I_s=I_s)
+
         V_ode = beat.utils.space_from_string(odespace, mesh, dim=1)
         v_ode = dolfinx.fem.Function(V_ode)
 
-        pde = beat.MonodomainModel(time=time, mesh=mesh, M=M, I_s=I_s, v_ode=v_ode)
         s = dolfinx.fem.Function(V_ode)
         s.interpolate(
             dolfinx.fem.Expression(s_exact_func(x, time), V_ode.element.interpolation_points()),
@@ -123,6 +125,7 @@ def test_monodomain_splitting_spatial_convergence(odespace):
 
         ode = beat.odesolver.DolfinODESolver(
             v_ode=v_ode,
+            v_pde=pde.state,
             fun=simple_ode_forward_euler,
             init_states=init_states,
             parameters=None,
@@ -162,6 +165,7 @@ def test_monodomain_splitting_temporal_convergence(theta, odespace):
     N = 150
     mesh = dolfinx.mesh.create_unit_square(comm, N, N, dolfinx.cpp.mesh.CellType.triangle)
     V_ode = beat.utils.space_from_string(odespace, mesh, dim=1)
+    v_ode = dolfinx.fem.Function(V_ode)
 
     errors = []
 
@@ -171,8 +175,8 @@ def test_monodomain_splitting_temporal_convergence(theta, odespace):
         x = ufl.SpatialCoordinate(mesh)
 
         I_s = ac_func(x, time)
-        v_ode = dolfinx.fem.Function(V_ode)
-        pde = beat.MonodomainModel(time=time, mesh=mesh, M=M, I_s=I_s, v_ode=v_ode)
+
+        pde = beat.MonodomainModel(time=time, mesh=mesh, M=M, I_s=I_s)
 
         s = dolfinx.fem.Function(V_ode)
         s.interpolate(
@@ -185,6 +189,7 @@ def test_monodomain_splitting_temporal_convergence(theta, odespace):
 
         ode = beat.odesolver.DolfinODESolver(
             v_ode=v_ode,
+            v_pde=pde.state,
             fun=simple_ode_forward_euler,
             init_states=init_states,
             parameters=None,
