### 2025-05-26 35de6c3
Revert "Revert "Do not transfer values between ODE and PDE, but rather use ODE directly""
