### 2025-05-26 9711623
Start implementing missing variables
