### 2025-06-08 6097a1f
Start working on cli for running simulations
