+++ src/beat/config.py
@@ -1,0 +1,191 @@
+import logging
+from pathlib import Path
+from typing import Annotated
+
+try:
+    from pydantic_pint import PydanticPintQuantity
+except ImportError:
+    msg = (
+        "pydantic_pint is not installed. "
+        "Install it with 'pip install pydantic-pint' to use PydanticPintQuantity.",
+    )
+    logging.warning(msg)
+    import sys
+
+    sys.exit(1)
+
+from pint import Quantity
+from pydantic import AfterValidator, Field
+from pydantic_settings import BaseSettings
+
+logger = logging.getLogger(__name__)
+
+
+class Conductivity(BaseSettings):
+    sigma_el: Annotated[Quantity, PydanticPintQuantity("S/m")] = Field(
+        "0.62 S/m",
+        description="Extracellular conductivity in the longitudinal direction (S/m)",
+    )
+    sigma_et: Annotated[Quantity, PydanticPintQuantity("S/m")] = Field(
+        "0.24 S/m",
+        description="Extracellular conductivity in the transverse direction (S/m)",
+    )
+    sigma_il: Annotated[Quantity, PydanticPintQuantity("S/m")] = Field(
+        "0.17 S/m",
+        description="Intracellular conductivity in the longitudinal direction (S/m)",
+    )
+    sigma_it: Annotated[Quantity, PydanticPintQuantity("S/m")] = Field(
+        "0.019 S/m",
+        description="Intracellular conductivity in the transverse direction (S/m)",
+    )
+
+
+class EPConfig(BaseSettings):
+    conductivity: Conductivity = Conductivity()
+    chi: Annotated[Quantity, PydanticPintQuantity("cm**-1")] = Field(
+        "1400 cm**-1",
+        description="Surface to volume ratio (cm^-1)",
+    )
+    C_m: Annotated[Quantity, PydanticPintQuantity("uF/mm**2")] = Field(
+        "0.01 uF/mm**2",
+        description="Membrane capacitance (uF/mm^2)",
+    )
+
+
+class MeshConfig(BaseSettings):
+    unit: str = Field("mm", description="Unit of the mesh")
+    folder: Path = Field("mesh", description="Folder containing the mesh files")
+
+
+def check_file_exists(file_path: Path) -> Path:
+    """
+    Ensure the specified file exists.
+
+    Parameters
+    ----------
+    file_path : Path
+        The path to the file to check.
+
+    Returns
+    -------
+    Path
+        The validated file path.
+
+    Raises
+    ------
+    FileNotFoundError
+        If the file does not exist.
+    """
+    file_path = file_path.resolve()
+    if not file_path.exists():
+        logger.warning(f"File not found: {file_path}")
+    return file_path
+
+
+class CellConfig(BaseSettings):
+    ode_file: Annotated[Path, AfterValidator(check_file_exists)] = Field(
+        "model.ode",
+        description="Path to the .ode model",
+    )
+    num_beats: int = Field(10, description="Number of beats for the single cell model")
+    BCL: Annotated[Quantity, PydanticPintQuantity("ms")] = Field(
+        "1000 ms",
+        description="Basic cycle length for the single cell simulation (ms)",
+    )
+    dt: Annotated[Quantity, PydanticPintQuantity("ms")] = Field(
+        "0.01 ms",
+        description="Time step for single cell simulations (ms)",
+    )
+    module_name: Path = Field(
+        "cell_model.py", description="Python module containing the cell model",
+    )
+    scheme: str = Field("generalized_rush_larsen", description="Scheme for the cell model")
+    track_indices: list[str] = Field(
+        default_factory=lambda: ["v", "cai"],
+        description="List of state indices to track during simulation",
+    )
+
+
[…]
