### 2025-06-08 e9a0172
Start working on cli for running simulations
