+++ .cspell_dict.txt
@@ -130,6 +130,7 @@ pmatrix
 Potse
 preconditioner
 preonly
+prepace
 pyplot
 PYVISTA
 Quarteroni
@@ -171,6 +172,7 @@ Virag
 viscoelastic
 viscoelasticity
 Waldman
+WRONLY
 XDMF
 Xiaodong
 xlabel
