### 2025-06-10 #8215 @cudmore
Same as @mgalloy, After looking through the paper and repo, I don't think I have much expertise in the area of this submission.

I am more of a cell biologist for electrophysiology and imaging. A bit too 'system' for me...
