### 2025-06-17 #8416 @editorialbot
[REVIEW]: fenicsx-beat - An Open Source Simulation Framework for Cardiac Electrophysiology
**Submitting author:** @finsberg (<a href="http://orcid.org/0000-0003-3766-2393">Henrik Nicolay Topnes Finsberg</a>)
**Repository:** https://github.com/finsberg/fenicsx-beat
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.3
