### 2025-07-15 #8416 @mbarzegary
Hi @sitic, I see you have opened an issue on the software repo regarding the paper. How is your review going regarding the software functionality?
