### 2025-07-21 a60682a
Bump version: 0.2.1 → 0.2.2
