### 2025-08-09 #8416 @ashahide
Hi @mbarzegary and @finsberg

Sorry for the delay on my end - I have completed my initial review. The software is interesting, well-constructed, and overall satisfactorily documented.

The documentation is missing a statement of need, which is a required item on the reviewer checklist.

Outside of the necessary changes, but as a suggestion, I would update the pip installation instructions. In order to install the optional dependencies required for the provided demo files, the user should run something like

`pip install .[all]`

as running

`pip install .`

installs only the required dependencies. While people familiar with pip and pyproject.toml will probably know this, documentation can make it more clear.
