### 2025-08-10 4efae46
Add statement of need and updated installation inststructions in the readme
