+++ .github/workflows/main-mpi.yml
@@ -16,7 +16,7 @@ jobs:
 
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v5
 
       - name: Install package
         run: |
