+++ .github/workflows/build_docs.yml
@@ -42,6 +42,6 @@ jobs:
         run: jupyter book build -W --keep-going .
 
       - name: Upload artifact
-        uses: actions/upload-pages-artifact@v3
+        uses: actions/upload-pages-artifact@v4
         with:
           path: ${{ env.PUBLISH_DIR }}
