### 2025-09-17 #8416 @mbarzegary
@finsberg given the green light of the reviewers, we will now work towards processing this for acceptance in JOSS. So please
* Merge my PR with very minor edits
* Work on the author's points of the final checklist I created ☝️
