### 2025-09-17 f020a88
Bump version: 0.2.2 → 0.2.3
