### 2025-09-18 #8416 @mbarzegary
@finsberg Please make sure that the title and author list of the archive match the JOSS paper. The title is wrong and the author list should contain you only.
