+++ joss-paper/paper.md
@@ -52,210 +52,6 @@ where $v$ is the transmembrane potential, $s$ represents the state variables of 
 
 The software leverages the FEniCSx library for the spatial discretization of the PDE component using the finite element method. Variational forms are expressed using the Unified Form Language (UFL), and the software utilizes the high-performance assembly and solution capabilities of DOLFINx. This allows for simulations on complex geometries using various element types and supports parallel execution via MPI. DOLFINx interfaces with robust external libraries, notably PETSc, for scalable linear algebra operations and solvers.
 
-
-## Illustrative example
-The following minimal example demonstrates simulating the Monodomain model on a unit square domain using a modified FitzHugh-Nagumo model [@rogers1994collocation]. For simplicity, we use the forward Euler method to solve the ODEs. However, in a real simulation, we would use `gotranx` [@Finsberg2024] to generate code for solving the ODEs (see more examples in the documentation).
-
-The state variables are $s$ and $v$. The derivatives $\frac{ds}{dt}$ and $\frac{dv}{dt}$ are calculated, and then the forward Euler update rule is applied.
-
-The equations for the derivatives are:
-
-$$ \frac{ds}{dt} = b((v - v_{rest})-c_3 s) $$
-$$ \frac{dv}{dt} = I + i_{app} $$
-
-where
-
-$$ I =  \left(\left(\frac{c_1}{v_{\mathrm{amp}}^2}\right) (v - v_{\mathrm{res}t})\right) (v - v_{th}) (v_{\mathrm{peak}} - v) - s \left(\frac{c_2}{v_{\mathrm{amp}}}\right) (v - v_{\mathrm{rest}})
-$$
-
-and the threshold voltage $v_{th}$ is defined as:
-
-$$ v_{th} = v_{\mathrm{amp}} a + v_{rest} $$
-
-The applied current $i_{app}(t)$ is defined as:
-
-$$
-i_{\mathrm{app}}(t) = \begin{cases} \mathrm{stim}_{\mathrm{amp}} & \text{if } \mathrm{stim}_{\mathrm{start}} < t < \mathrm{stim}_{\mathrm{start}} + \mathrm{stim}_{\mathrm{dur}} \\ 0 & \text{otherwise} \end{cases}.
-$$
-
-The following code snippet demonstrates the setup and execution of a simulation using `fenicsx-beat`:
-
-
-```python
-import shutil
-
-import matplotlib.pyplot as plt
-import numpy as np
-
-from mpi4py import MPI
-import dolfinx
-import ufl
-
-import beat
-
-# MPI communicator
-comm = MPI.COMM_WORLD
-# Create mesh
-mesh = dolfinx.mesh.create_unit_square(comm, 32, 32, dolfinx.cpp.mesh.CellType.triangle)
-# Create a variable for time
-time = dolfinx.fem.Constant(mesh, dolfinx.default_scalar_type(0.0))
-
-
-# Define forward euler scheme for solving the ODEs
-# This just needs to be a function that takes the current time, states, parameters and dt
-# and returns the new states
-def fitzhughnagumo_forward_euler(t, states, parameters, dt):
-    s, v = states
-    (
-        c_1,
-        c_2,
-        c_3,
-        a,
-        b,
-        v_amp,
-        v_rest,
-        v_peak,
-        stim_amplitude,
-        stim_duration,
-        stim_start,
-    ) = parameters
-    i_app = np.where(
-        np.logical_and(t > stim_start, t < stim_start + stim_duration),
-        stim_amplitude,
-        0,
[…]
