+++ joss-paper/paper.md
@@ -50,7 +50,7 @@ $$
 
 where $v$ is the transmembrane potential, $s$ represents the state variables of the ionic model, $C_m$ is the membrane capacitance, $I_{\text{ion}}$ is the total ionic current, $I_{\text{stim}}$ is the applied stimulus current, $\sigma$ is the conductivity tensor, and $f(s, v, t)$ is the system of ODEs defining the ionic model. The Monodomain equation describes the diffusion of the transmembrane potential $v$ in the tissue, while the ODE system captures the dynamics of the ionic currents across the cell membrane. The coupling between these two components is essential for accurately simulating cardiac electrophysiology.
 
-The software leverages the FEniCSx library for the spatial discretization of the PDE component using the finite element method. Variational forms are expressed using the Unified Form Language (UFL), and the software utilizes the high-performance assembly and solution capabilities of DOLFINx. This allows for simulations on complex geometries using various element types and supports parallel execution via MPI. DOLFINx interfaces with robust external libraries, notably PETSc, for scalable linear algebra operations and solvers.
+The software leverages the FEniCSx library for the spatial discretization of the PDE component using the finite element method. Variational forms are expressed using the Unified Form Language (UFL), and the software utilizes the high-performance assembly and solution capabilities of DOLFINx. This allows for simulations on complex geometries using various element types and supports parallel execution via MPI. DOLFINx interfaces with robust external libraries, notably PETSc[@petsc-user-ref], for scalable linear algebra operations and solvers.
 
 ## Comparison with Other Software
 The field of computational cardiac electrophysiology benefits from several open-source simulation packages. `fenicsx-beat` distinguishes itself by being built natively on the modern FEniCSx framework, targeting users who leverage this platform for its flexibility in solving PDEs with the finite element method.
