+++ joss-paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 
 # Statement of need
-Computational modeling plays an increasingly vital role in cardiac electrophysiology research, offering insights into mechanisms of heart rhythm disorders (arrhythmias), evaluating therapeutic strategies like drug effects or ablation, and paving the way towards personalized medicine and 'digital twin' concepts. Mathematical models, ranging from detailed ionic interactions at the subcellular level to electrical wave propagation at the tissue and organ level, allow for quantitative investigation of physiological and pathological conditions that are difficult or impossible to study experimentally.
+Computational modeling plays an increasingly vital role in cardiac electrophysiology research, offering insights into mechanisms of heart rhythm disorders (arrhythmias), evaluating therapeutic strategies like drug effects or ablation, and paving the way towards personalized medicine and "digital twin" concepts. Mathematical models, ranging from detailed ionic interactions at the subcellular level to electrical wave propagation at the tissue and organ level, allow for quantitative investigation of physiological and pathological conditions that are difficult or impossible to study experimentally.
 
 The Bidomain model is often considered the most physiologically detailed representation of cardiac tissue electrophysiology, but its computational cost can be prohibitive for large-scale or long-duration simulations. The Monodomain model, a simplification valid under certain assumptions about tissue conductivity, offers a computationally less expensive alternative while still capturing essential propagation dynamics accurately in many scenarios, such as studies of activation sequences or basic arrhythmia mechanisms. Solving these models typically involves coupling the reaction-diffusion PDE (Monodomain or Bidomain) with a system of stiff ODEs describing the ionic currents across the cell membrane (the ionic model) [@sundnes2007computing].
 
@@ -55,7 +55,7 @@ The software leverages the FEniCSx library for the spatial discretization of the
 ## Comparison with Other Software
 The field of computational cardiac electrophysiology benefits from several open-source simulation packages. `fenicsx-beat` distinguishes itself by being built natively on the modern FEniCSx framework, targeting users who leverage this platform for its flexibility in solving PDEs with the finite element method.
 
-Within the FEniCS ecosystem, `fenicsx-beat` can be seen as a successor or counterpart to `cbcbeat`, which provided similar Monodomain/Bidomain capabilities but was based on the legacy FEniCS library. Other FEniCS/FEniCSx-based tools focus on different physics: `simcardems` couples electrophysiology with solid mechanics, `pulse` [@Finsberg2019] focuses solely on cardiac mechanics, and `Ambit` [@Hirschvogel2024] is a newer multi-physics solver primarily for cardiac mechanics and fluid-structure interaction (FSI), although future electrophysiology capabilities are envisioned. `fenicsx-beat` provides the dedicated, up-to-date electrophysiology solver within this FEniCSx environment.
+Within the FEniCS ecosystem, `fenicsx-beat` can be seen as a successor or counterpart to `cbcbeat`, which provided similar Monodomain/Bidomain capabilities but was based on the legacy FEniCS library. Other FEniCS/FEniCSx-based tools focus on different physics: `simcardems` couples electrophysiology with solid mechanics, `pulse` [@Finsberg2019] focusing solely on cardiac mechanics, and `Ambit` [@Hirschvogel2024] is a newer multi-physics solver primarily for cardiac mechanics and fluid-structure interaction (FSI), although future electrophysiology capabilities are envisioned. `fenicsx-beat` provides a dedicated, up-to-date electrophysiology solver within this FEniCSx environment.
 
 Compared to established standalone simulators like openCARP [@plank2021opencarp], `fenicsx-beat` offers tighter integration with FEniCSx libraries. openCARP is a powerful, widely used simulator with its own optimized C++ core and a dedicated Python workflow tool (carputils) for managing complex simulations.
 
