+++ joss-paper/paper.md
@@ -57,9 +57,7 @@ The field of computational cardiac electrophysiology benefits from several open-
 
 Within the FEniCS ecosystem, `fenicsx-beat` can be seen as a successor or counterpart to `cbcbeat`, which provided similar Monodomain/Bidomain capabilities but was based on the legacy FEniCS library. Other FEniCS/FEniCSx-based tools focus on different physics: `simcardems` couples electrophysiology with solid mechanics, `pulse` [@Finsberg2019] focusing solely on cardiac mechanics, and `Ambit` [@Hirschvogel2024] is a newer multi-physics solver primarily for cardiac mechanics and fluid-structure interaction (FSI), although future electrophysiology capabilities are envisioned. `fenicsx-beat` provides a dedicated, up-to-date electrophysiology solver within this FEniCSx environment.
 
-Compared to established standalone simulators like openCARP [@plank2021opencarp], `fenicsx-beat` offers tighter integration with FEniCSx libraries. openCARP is a powerful, widely used simulator with its own optimized C++ core and a dedicated Python workflow tool (carputils) for managing complex simulations.
-
-Another prominent open-source package is Chaste (Cancer, Heart And Soft Tissue Environment) [@Cooper2020], a C++ library designed for a broad range of computational biology problems, including cardiac electrophysiology.
+Compared to established standalone simulators like openCARP [@plank2021opencarp] or Chaste[@Cooper2020], `fenicsx-beat` offers tighter integration with FEniCSx libraries. openCARP is a powerful, widely used simulator with its own optimized C++ core and a dedicated Python workflow tool (carputils) for managing complex simulations. Chaste (Cancer, Heart And Soft Tissue Environment) is a C++ library designed for a broad range of computational biology problems, including cardiac electrophysiology.
 
 In summary, `fenicsx-beat` occupies a valuable niche by providing a modern, Python-interfaced, FEniCSx-native tool for Monodomain cardiac electrophysiology simulations. Its primary strength lies in its seamless integration with the FEniCSx platform, making it an attractive choice for researchers utilizing FEniCSx for multi-physics cardiac modeling.
 
