+++ pyproject.toml
@@ -6,7 +6,6 @@ name = "fenicsx-beat"
 version = "0.2.3"
 description = "Library to run cardiac EP simulations"
 authors = [{name = "Henrik Finsberg", email = "henriknf@simula.no"}]
-license = "MIT"
 license-files = ["LICENSE"]
 readme = "README.md"
 requires-python = ">=3.8"
