### 2025-10-05 662eef8
Add scikit-build-core build deps
