+++ pyproject.toml
@@ -1,11 +1,12 @@
 [build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
-requires = ["setuptools>=61.0.0", "wheel"]
+requires = ["setuptools>=61.0.0", "wheel", "scikit-build-core[pyproject]>=0.11.0"]
 
 [project]
 name = "fenicsx-beat"
 version = "0.2.3"
 description = "Library to run cardiac EP simulations"
 authors = [{name = "Henrik Finsberg", email = "henriknf@simula.no"}]
+license = "MIT"
 license-files = ["LICENSE"]
 readme = "README.md"
 requires-python = ">=3.8"
