### 2025-10-05 874679c
Add more build dependencies
