+++ pyproject.toml
@@ -1,5 +1,11 @@
 [build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
-requires = ["setuptools>=61.0.0", "wheel", "scikit-build-core[pyproject]>=0.11.0"]
+requires = [
+    "setuptools>=77.0.3",                   # Ensuring license field is correctly parsed]
+    "packaging>=24.2",                      # Ensuring license field is correctly parsed]
+    "pkgconfig",
+    "wheel",
+    "scikit-build-core[pyproject]>=0.11.0"
+]
 
 [project]
 name = "fenicsx-beat"
