### 2025-10-05 90dc48a
Install build dependencies from scifem
