### 2025-10-05 #8416 @finsberg
Hi @rkurchin thanks for the review. I belive I have now addressed all the points in https://github.com/finsberg/fenicsx-beat/pull/33
