### 2025-10-06 dc833aa
Require pyvista>=0.45.0
