+++ pyproject.toml
@@ -57,6 +57,8 @@ docs = [
    "fenicsx-beat[demos]",
    "ipywidgets",
    "imageio",
+   "pyvista[all]>=0.45.0",
+   "vtk>=9.5.0",
    "pyvista[jupyter]",
    "sphinx-codeautolink",
    "sphinxcontrib-bibtex",
