### 2025-10-08 49be9b5
Minor edits to joss paper
