### 2025-10-08 #8416 @finsberg
No worries :) Fixed in https://github.com/finsberg/fenicsx-beat/commit/49be9b5f22d12071fa1e86f358fbbafffec89b07
